requirements_base.txt
dj-database-url==0.4.1
dj-static==0.0.6
Django==1.8.8
django-countries==4.0
django-easymoney==0.7.1
django-floppyforms==1.7.0
django-idmap==0.4.1
django-toolbelt==0.0.1
django-vanilla-views==1.0.4
djangorestframework==3.4.6
otree-core==1.2.7
otree-save-the-change==1.1.3